% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_diagram.R
\name{plot_tree_diagram}
\alias{plot_tree_diagram}
\title{Plot a tree diagram}
\usage{
plot_tree_diagram(
  df,
  root_name,
  tree_direction = "LR",
  remove_0 = TRUE,
  edge_p_digits = 3,
  joint_p_digits = 5,
  joint_p_scientific = TRUE,
  bgcolor = "lightgrey",
  joint_p_edge_style = ""
)
}
\arguments{
\item{df}{The input data frame. Each row should represent a leaf in the data.tree structure.
Also, the columns should be in order of level_1_var, Prob_level_1, level_2_var, Prob_level_2, ...
The 1st column should contain values of the level 1 variable;
the 2nd column should contain the probability of level 1 variable equal to the corresponding value in the same row of the 1st column;
the 3rd column should contain values of the level 2 variable;
the 4nd column should contain the conditional probability of level 2 variable given the level 1 variable}

\item{root_name}{a character string specifying the name of the root node}

\item{tree_direction}{a character string specifying the direction of the tree diagram. The default is "LR", meaning
from left to right. It can also be "TB" (top to bottom), "BT" (bottom to top), "RT"(right to left).}

\item{remove_0}{Logical. Set to TRUE if we don't want to show leaves/paths with joint probability=0.}

\item{edge_p_digits}{Positive integer indicating the number of decimal places to be}

\item{joint_p_digits}{positive integer indicating the number of decimal places to be
used when rounding joint probabilities. The default value is 5.}

\item{joint_p_scientific}{Logical. Specify whether the joint probabilities should be
displayed in scientific format. The default value is TRUE.}

\item{bgcolor}{Specify background color.}

\item{joint_p_edge_style}{can set to "invis" so that there is no edge connecting the leaves and the joint prob.}
}
\value{
a tree diagram
}
\description{
Plot a tree diagram
}
\examples{
#plot for Monty_hall
data(Monty_hall)
plot_tree_diagram(Monty_hall, root_name="Monty Hall")
plot_tree_diagram(Monty_hall, root_name="Monty Hall", remove_0 = FALSE)

#plot for X~Poisson(3)
test_df <- prepare_df(X_dist="Pois", X_parameters= c(3))
plot_tree_diagram(test_df,root_name="X")

#plot for X~NegBin(2,0.6), Y|X ~ Bern(1/X)
test_df <- prepare_df(X_dist="NegBin", X_parameters= c(2,0.6),
                      Y_given_X_dist="Bin", Y_parameters = c(1,"1/X"))
plot_tree_diagram(test_df, root_name="X")
}
