% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_df.R
\name{prepare_df}
\alias{prepare_df}
\title{Prepare the input data frame for plot_tree_diagram()}
\usage{
prepare_df(
  X_dist,
  X_parameters,
  X_low_q = 0.01,
  X_high_q = 0.99,
  Y_given_X_dist = NULL,
  Y_parameters,
  Y_low_q = 0.1,
  Y_high_q = 0.9
)
}
\arguments{
\item{X_dist}{Specify distribution of level 1 variable. X_dist can be "Pois",
"Geo", "Bin", or "NegBin". If you want X_dist to be Bernoulli distribution,
please specify X_dist to be "Bin" because Bernoulli is just
a special case of Binomial distribution.}

\item{X_parameters}{a vector containing the parameters for X_dist.}

\item{X_low_q}{the left quantile of X. Used to determine the minimum value of
X to be included in the tree diagram.}

\item{X_high_q}{the right quantile of X. Used to determine the maximum value of
X to be included in the tree diagram.}

\item{Y_given_X_dist}{Specify distribution of level 2 variable. Y_given_X_dist
can be "Pois", "Geo", "Bin", or "NegBin". If you want Y_given_X_dist to be
Bernoulli distribution, please specify Y_given_X_dist to be "Bin" because Bernoulli
is just a special case of Binomial distribution.}

\item{Y_parameters}{a vector containing parameter(s) for Y_given_X_dist.
Can use a string to express the relationship between any parameter and the level 1 variable X.
For example, if Y|X ~ Bern(1/X), then Y_given_X_dist = "Bin", Y_parameters = c(1,"1/X");}

\item{Y_low_q}{the left quantile of Y|X=x. Used to determine the minimum value of
Y to be included in the tree diagram given a specific value of X.}

\item{Y_high_q}{the right quantile of Y|X=x. Used to determine the maximum value of
Y to be included in the tree diagram given a specific value of X.}
}
\value{
a data frame that can be used as the input df to the plot_tree_diagram().
}
\description{
Prepare the input data frame for plot_tree_diagram()
}
\examples{
#For X~Poisson(3)
test_df <- prepare_df(X_dist="Pois", X_parameters= c(3))

#for X~NegBin(2,0.6), Y|X ~ Bern(1/X)
test_df <- prepare_df(X_dist="NegBin", X_parameters= c(2,0.6),
                      Y_given_X_dist="Bin", Y_parameters = c(1,"1/X"))

#for X~Poisson(2), Y|X ~ Bin(X,1/X)
test_df <- prepare_df(X_dist="Pois", X_parameters= c(2),
          Y_given_X_dist="Bin", Y_parameters = c("X","1/X"))
}
